Server now has a much better look, and a dedicated log method. The server also no longer will send a message to a client who is connected but does not have an establised connection. The client now can send messages, set their username, and the such.
Because of how the server requires a message before it will allow the user to send a nickname, you can use this as a password lock of some kind on custom servers. When JSON is implemented, this will allow servers to specify if a client should send a password or not.