Added a few neat things to the client and the server. First, the client is now has a state. This state indicates what the next message can indicate, for instance: if a password was correct.
The client now has a better look, you can now see the seperation between messages, different log types, ect.

The server can now tell the client that is connecting if they must send a password to join. If it is required, the user will be prompted to send one, otherwise it will say that it is connecting.

User connects and disconnects are now broadcasted to all clients.